pyproject.toml
[project]
name = "Welead"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "django>=5.1.7",
    "django-allauth>=65.4.1",
    "django-cors-headers>=4.7.0",
    "django-filter>=25.1",
    "django-jazzmin>=3.0.1",
    "django-split-settings>=1.3.2",
    "djangorestframework>=3.15.2",
    "djangorestframework-simplejwt>=5.5.0",
    "pillow>=11.1.0",
    "pyjwt>=2.9.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
]

[dependency-groups]
dev = [
    "daphne>=4.1.2",
    "django-debug-toolbar>=5.0.1",
    "icecream>=2.1.4",
    "pytest>=8.3.5",
]
production = [
    "whitenoise>=6.9.0",
]
